@@ -5,7 +5,7 @@
   "main": "src/index.ts",
   "scripts": {
     "lint": "eslint .",
-    "build": "tsc -d",
+    "build": "webpack --config build/webpack.config.js",
     "lint:style": "stylelint \"**/*.{css,scss}\" --fix",
     "lint:fix": "eslint . --ext \".ts,.tsx\" --fix",
     "test": "echo \"Error: no test specified\" && exit 1",
@@ -27,7 +27,8 @@
     "lodash": "^4.17.21",
     "lunar-typescript": "^1.2.14",
     "number-precision": "^1.5.1",
-    "react": "^17.0.2",
+    "react": "^18.0.0-rc.2-next-581f0c42e-20220310",
+    "react-dom": "^18.0.0-rc.2-next-581f0c42e-20220310",
     "recoil": "^0.6.1",
     "rxjs": "^7.4.0",
     "standard-version": "^9.3.1",
@@ -57,6 +58,7 @@
     "eslint-plugin-react-hooks": "^4.2.0",
     "husky": "^7.0.1",
     "npm": "^7.20.6",
+    "postcss-loader": "^6.2.1",
     "prettier": "2.3.2",
     "sass": "^1.37.5",
     "stylelint": "^13.13.1",
