@@ -27,8 +27,10 @@
     "highlight.js": "^11.4.0",
     "lz-string": "^1.4.4",
     "monaco-editor": "^0.32.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "monaco-editor-nls": "^2.0.0",
+    "react": "^18.0.0-rc.2-next-581f0c42e-20220310",
+    "react-dom": "^18.0.0-rc.2-next-581f0c42e-20220310",
+    "react-helmet": "^6.1.0",
     "react-live": "^2.4.1",
     "react-router-dom": "5"
   },
@@ -81,6 +83,7 @@
     "markdown-it-container": "^3.0.0",
     "markdown-it-toc-done-right": "^4.2.0",
     "mini-css-extract-plugin": "^2.5.3",
+    "monaco-editor-esm-webpack-plugin": "^2.0.2",
     "monaco-editor-webpack-plugin": "^7.0.1",
     "optimize-css-assets-webpack-plugin": "^6.0.1",
     "portfinder": "^1.0.28",
@@ -101,6 +104,7 @@
     "webpack-bundle-analyzer": "^4.5.0",
     "webpack-cli": "^4.9.2",
     "webpack-dev-server": "^4.7.4",
-    "webpack-merge": "^5.8.0"
+    "webpack-merge": "^5.8.0",
+    "worker-loader": "^3.0.8"
   }
 }
