@@ -1,11 +1,20 @@
-import React, {useRef, useState, useEffect} from 'react'
+
+
+
 import * as ts  from 'typescript'
+
+import React, {useRef, useState, useEffect } from 'react'
 import { createSystem, createDefaultMapFromCDN, createVirtualCompilerHost } from '@typescript/vfs'
 import lzstring from 'lz-string'
 import { Sandbox } from '../sandbox/sandbox'
-import * as monaco from 'monaco-editor'
+import { setLocaleData } from 'monaco-editor-nls'
+import zh_CN from 'monaco-editor-nls/locale/zh-hans.json'
+
+setLocaleData(zh_CN)
 
 import './style.scss'
+import * as monaco from 'monaco-editor/esm/vs/editor/editor.api'
+// const monaco = require('monaco-editor');
 
 interface EditorProps {
   code: string
@@ -19,79 +28,81 @@ export const Editor = (props: EditorProps) => {
   } = props
   const divEl = useRef<HTMLDivElement>(null)
   const [transformCode, setTransformCode] = useState('')
+  const compilerOptions = {
+    noEmit: false,
+    noEmitOnError: true,
+    module: ts.ModuleKind.CommonJS, 
+    target: ts.ScriptTarget.ES5,
+    jsx: ts.JsxEmit.React,
+    moduleResolution: ts.ModuleResolutionKind.NodeJs
+  }
 
   let editor: monaco.editor.IStandaloneCodeEditor
 
-  // ts-ignore
-  useEffect(async () => {
-    if (divEl.current) {
-      const compilerOptions = {
-        // esnext
-        module: 99, 
-        // es5
-        target: 1,
-        // react
-        jsx: 2,
-        // node
-        moduleResolution: 2
-      }
+ 
 
-      const fsMap = await createDefaultMapFromCDN(compilerOptions, ts.version, false, ts, lzstring)
+  const initEditor = () => {
+    editor = monaco.editor.create(divEl.current, {
+      value: code,
+      language: 'typescript',
+      theme: 'vs',
+      wordWrap: 'on',
+      minimap: {
+        enabled: false
+      },
+    })
+  
+    monaco.languages.typescript.typescriptDefaults.setCompilerOptions({
+      jsx: ts.JsxEmit.React
+    })
 
-      fsMap.set('index.tsx', code)
-
-      const system = createSystem(fsMap)
-      const host = createVirtualCompilerHost(system, compilerOptions, ts)
-
-      const program = ts.createProgram({
-        rootNames: [...fsMap.keys()],
-        options: compilerOptions,
-        host: host.compilerHost
-      })
-
-      console.log(ts)
-      console.log(program.getTypeChecker(), [...fsMap.keys()])
-
-      // console.log(program.emitFile())
-      const demoCode = `
-        function Demo() {
-          const a: number = 1
-
-          return (
-            <div>hello world {a}</div>
-          )
-        }
-
-        ReactDOM.render(<Demo/>, container)
-      `
+    editor.onDidChangeModelContent(() => {
+      console.log(editor.getValue())
       setTransformCode(
-        ts.transpile(code, compilerOptions)
+        ts.transpile(editor.getValue(), compilerOptions)
       )
-      console.log(ts.transpile(code, compilerOptions))
-      // env.languageService.getDocumentHighlights('index.tsx', 0, ['index.tsx'])
+    })
+  }
 
+  useEffect(() => {
+    setTransformCode(
+      ts.transpile(code, compilerOptions)
+    )
+  
+    console.log(transformCode)
 
-      editor = monaco.editor.create(divEl.current, {
-        value: code,
-        language: 'typescript',
-        theme: 'vs'
-      })
+    const getTypescriptCDN = async () => {
+      if (divEl.current) {
+        initEditor()
 
-      monaco.languages.typescript.typescriptDefaults.setCompilerOptions({
-        jsx: 2
-      })
-
-      editor.onDidChangeModelContent(e => {
-        setTransformCode(
-          ts.transpile(editor.getValue(), compilerOptions)
-        )
-        console.log(e, editor.getValue())
-      })
+  
+        // const fsMap = await createDefaultMapFromCDN(compilerOptions, ts.version, false, ts, lzstring)
+  
+        // fsMap.set('index.tsx', code)
+  
+        // const system = createSystem(fsMap)
+        // const host = createVirtualCompilerHost(system, compilerOptions, ts)
+  
+        // const program = ts.createProgram({
+        //   rootNames: [...fsMap.keys()],
+        //   options: compilerOptions,
+        //   host: host.compilerHost
+        // })
+        
+        // env.languageService.getDocumentHighlights('index.tsx', 0, ['index.tsx'])
+  
+  
+     
+      }
     }
+
+    getTypescriptCDN()
+    
     return () => {
-      editor.dispose()
+      editor?.dispose()
     }
   }, [])
+
 
   return  (
     <>
